mariadb: Fix networking issues
The mariadb package for alpine defaults skip-networking to 1
in order to allow networking the value has to be explicitly set to 0.

Since this has to be run in a container has to listen on all addresses,
this is not dangerous since the port has to be published to the host.